Fix ext32 unpacking of the maximum UINT32_MAX-byte payload
The C unpacker's internal "trail" bookkeeping combines an ext value's
type byte with its data into a single byte count by adding 1 to the
length read off the wire. For ext8/ext16 that addition stays well
inside a 32-bit range, but ext32's length field can legitimately be
the full UINT32_MAX, and tmp + 1 done in 32-bit arithmetic wraps
around to 0. The parser then treats the object as a zero-length ext
and immediately errors out through template_callback_ext's l == 0
check, so any ext exactly at the wire format's own maximum size fails
to unpack with MSGPACK_UNPACK_PARSE_ERROR even though it's a
perfectly valid encoding.

Widened the trail field (and the str/bin/ext callback length
parameter that ends up receiving it) from unsigned int to size_t, and
cast the ext32 length to size_t before adding 1 so the addition
itself doesn't wrap on a 64-bit build. On a 32-bit build the same
edge case is still out of reach, but that matches the existing
intentional limitation on the C++ side and a 32-bit process can't
address a buffer that size anyway.

Added a regression test that packs and unpacks an ext with exactly
UINT32_MAX bytes of data; it skips itself if the host can't spare the
roughly 8GB of memory the test needs rather than failing outright.

Fixes #1086

+++ include/msgpack/unpack_template.h
@@ -53,7 +53,7 @@ msgpack_unpack_struct_decl(_stack) {
 msgpack_unpack_struct_decl(_context) {
     msgpack_unpack_user user;
     unsigned int cs;
-    unsigned int trail;
+    size_t trail;
     unsigned int top;
     /*
     msgpack_unpack_struct(_stack)* stack;
@@ -99,7 +99,7 @@ msgpack_unpack_func(int, _execute)(msgpack_unpack_struct(_context)* ctx, const c
         const unsigned char* const pe = (unsigned char*)data + len;
         const void* n = NULL;
 
-        unsigned int trail = ctx->trail;
+        size_t trail = ctx->trail;
         unsigned int cs = ctx->cs;
         unsigned int top = ctx->top;
         msgpack_unpack_struct(_stack)* stack = ctx->stack;
@@ -326,7 +326,7 @@ msgpack_unpack_func(int, _execute)(msgpack_unpack_struct(_context)* ctx, const c
                 case MSGPACK_CS_EXT_16:{
                     uint16_t tmp;
                     _msgpack_load16(uint16_t,n,&tmp);
-                    again_fixed_trail_if_zero(MSGPACK_ACS_EXT_VALUE, tmp + 1, _ext_zero);
+                    again_fixed_trail_if_zero(MSGPACK_ACS_EXT_VALUE, (size_t)tmp + 1, _ext_zero);
                 }
                 case MSGPACK_CS_STR_32:{
                     uint32_t tmp;
@@ -341,7 +341,11 @@ msgpack_unpack_func(int, _execute)(msgpack_unpack_struct(_context)* ctx, const c
                 case MSGPACK_CS_EXT_32:{
                     uint32_t tmp;
                     _msgpack_load32(uint32_t,n,&tmp);
-                    again_fixed_trail_if_zero(MSGPACK_ACS_EXT_VALUE, tmp + 1, _ext_zero);
+                    /* cast before adding: on a 64-bit size_t this lets an ext32 with
+                     * the maximum UINT32_MAX-byte payload carry its trailing type
+                     * byte without wrapping trail back to 0 the way `tmp + 1` would
+                     * when tmp is exactly UINT32_MAX */
+                    again_fixed_trail_if_zero(MSGPACK_ACS_EXT_VALUE, (size_t)tmp + 1, _ext_zero);
                 }
                 case MSGPACK_ACS_STR_VALUE:
                 _str_zero:
